+++ cypress/integration/normal-script/product.feature
@@ -1,37 +1,33 @@
 Feature: Product
 
-Scenario: Buy a product Amazon
-Given A user opens the the product page with AMZ
-And A user validates all information
+Scenario: Validate Amazon product
+Given A user opens an Amazon product
+And A user validates product information
+And A user clicks the add to cart button
 And A user clicks the buy button
 Then The user will be on the cart page
+And The user validates there are 2 products in the cart
 
-Scenario: Add a product to my cart Amazon
-Given A user opens the the product page with AMZ
-And A user validates all information
-And A user clicks the add product button
-Then The user will see success label message
-
-Scenario: Buy a product Ebay
-Given A user opens the the product page with Ebay
-And A user validates all information
+Scenario: Validate eBay product
+Given A user opens an eBay product
+And A user validates product information
+And A user clicks the add to cart button
 And A user clicks the buy button
 Then The user will be on the cart page
+And The user validates there are 2 products in the cart
 
-Scenario: Add a product to my cart Ebay
-Given A user opens the the product page with Ebay
-And A user validates all information
-And A user clicks the add product button
-Then The user will see success label message
-
-Scenario: Buy a product Waltmart
-Given A user opens the the product page with WRT
-And A user validates all information
+Scenario: Validate Walmart product
+Given A user opens an Walmart product
+And A user validates product information
+And A user clicks the add to cart button
 And A user clicks the buy button
 Then The user will be on the cart page
+And The user validates there are 2 products in the cart
 
-Scenario: Add a product to my cart Waltmart
-Given A user opens the the product page with WRT
-And A user validates all information
-And A user clicks the add product button
-Then The user will see success label message
+Scenario: Validate Macy's product
+Given A user opens an Macy's product
+And A user validates product information
+And A user clicks the add to cart button
+And A user clicks the buy button
+Then The user will be on the cart page
+And The user validates there are 2 products in the cart
