+++ cypress/integration/normal-script/product/product.js
@@ -1,21 +1,38 @@
 import {Given, When, And, Then} from "cypress-cucumber-preprocessor/steps";
-import carritoPage from "../../../pages/carritoPage";
-const productPage = require('../../../pages/productPage')
+
+const carritoPage = require('../../../pages/carritoPage')
+const productPage = require('../../../pages/productPage');
+const amzUrl = '/product/amz/B000FZX93K';
+const ebyUrl = '/product/ebay/v1%7C264724079628%7C564540908530';
+const wrtUrl = '/product/wrt/633860434';
+const mcyUrl = '/product/mcy/8570453';
 
 Given('A user opens an Amazon product', ()=>{
-    cy.visit('/product/amz/B000FZX93K');
+    cy.visit(amzUrl).then(() => {
+        cy.url(amzUrl).should('include', amzUrl);
+        cy.wait(3000);
+    });
 })
 
 Given('A user opens an eBay product', ()=>{
-    cy.visit('/product/ebay/v1%7C264724079628%7C564540908530');
+    cy.visit(ebyUrl).then(() => {
+        cy.url(ebyUrl).should('include', ebyUrl);
+        cy.wait(3000);
+    });
 })
 
 Given('A user opens an Walmart product', ()=>{
-    cy.visit('/product/wrt/633860434');
+    cy.visit(wrtUrl).then(() => {
+        cy.url(wrtUrl).should('include', wrtUrl);
+        cy.wait(3000);
+    });
 })
 
 Given('A user opens an Macy\'s product', ()=>{
-    cy.visit('/product/mcy/8570453');
+    cy.visit(mcyUrl).then(() => {
+        cy.url(mcyUrl).should('include', mcyUrl);
+        cy.wait(3000);
+    });
 })
 
 When('A user validates product information', () => {
@@ -26,8 +43,13 @@ When('A user validates product information', () => {
     productPage.validateProductTitle();
 })
 
-And('A user clicks the buy button', () => {
+When('A user clicks the buy button', () => {
     productPage.clickBuyProductBtn();
+})
+
+When('A user clicks the add to cart button', () => {
+    productPage.clickAddToCartButton();
+    productPage.validateCartCounterLabel();
 })
 
 Then('The user will be on the cart page', () => {
@@ -42,6 +64,3 @@ Then('The user validates there are 2 products in the cart', () => {
     carritoPage.validateProductQuantity(2);
 })
 
-And('A user clicks the add to cart button', () => {
-    productPage.clickAddToCartButton();
-})
